+++ packages/tools/atlas/bin/dist_setup.sh
@@ -1,5 +1,0 @@
-#! /usr/bin/env bash
-
-rm -rf distro && \
-mkdir -p distro && \
-git clone git@github.com:nestauk/svizzle_atlas_distro.git distro