+++ packages/tools/atlas/src/bin/base/NUTS/run.sh
@@ -1,0 +1,11 @@
+#! /usr/bin/env bash
+
+echo '
+==== NUTS ====
+' &&
+
+node -r esm src/bin/base/NUTS/01_getMetadata.js &&
+node -r esm src/bin/base/NUTS/01_getShapes.js &&
+node -r esm src/bin/base/NUTS/12_getCountriesTopojson.js &&
+
+echo