+++ packages/tools/atlas/src/bin/base/runAll.sh
@@ -1,0 +1,6 @@
+#! /usr/bin/env bash
+
+# path are relative to the root of the package, where we run `runAll` from
+
+src/bin/base/NUTS/run.sh &&
+src/bin/base/world/run.sh