+++ packages/tools/atlas/src/bin/base/world/run.sh
@@ -1,0 +1,9 @@
+#! /usr/bin/env bash
+
+echo '
+==== world ====
+' &&
+
+node -r esm src/bin/base/world/01_world_110m_iso_a2_topojson.js &&
+
+echo